motor control: allow to disable zoom controls
The optical zoom controls and the use of the cell phone camera aren't
really practical for the astronomy observatins. ASI183MC astronomy
camera turned out to be an investment that made everything much easier
for the remote observing setup.
The resolution and pixel size of the camera are good enough to cover
everything a 12" Orion telescope can do.
The optical zoom control might still be useful for touch free adjusting
when observing directly or through PVS-14.